+++ crates/rolldown/src/bundler/graph/mod.rs
@@ -1,4 +1,5 @@
 #[allow(clippy::module_inception)]
 pub mod graph;
 pub mod linker;
+pub mod linker_info;
 pub mod symbols;
