+++ crates/rolldown/src/bundler/visitors/renderer_base.rs
@@ -4,9 +4,11 @@ use rolldown_oxc::BindingIdentifierExt;
 use rustc_hash::FxHashMap;
 use string_wizard::{MagicString, UpdateOptions};
 
+use crate::bundler::graph::linker_info::LinkingInfo;
+
 use super::super::{
   chunk_graph::ChunkGraph,
-  graph::{graph::Graph, linker::LinkingInfo},
+  graph::graph::Graph,
   module::{Module, NormalModule},
 };
 
@@ -88,8 +90,7 @@ impl<'ast> RendererBase<'ast> {
       let namespace_name = self.canonical_name_for(self.module.namespace_symbol);
       let exports: String = self
         .linking_info
-        .resolved_exports
-        .iter()
+        .exports()
         .map(|(exported_name, resolved_export)| {
           let canonical_ref = self.graph.symbols.par_canonical_ref_for(resolved_export.symbol_ref);
           let symbol = self.graph.symbols.get(canonical_ref);
